==> notebook/1 - MZ SCADA (MSSQL u CSV).ipynb <==
import sys
import pandas as pd
import sqlalchemy
from sqlalchemy import create_engine, MetaData, Table, select, engine

print('Python v' + sys.version)
print('------------------------------------------------------------------------------')
print('Pandas v' + pd.__version__)
print('SQLAlchemy v' + sqlalchemy.__version__)

DB = {
    'drivername': 'mssql+pyodbc',
    'servername': 'VERDI',
    'port': '1433',
    'username': 'nebojsa',
    'password': 'tras659',
    'database': 'SANiN',
    'driver': 'SQL Server Native Client 11.0',
#     'trusted_connection': 'yes',  
#     'legacy_schema_aliasing': 'False'
}

# conn_string = DB['drivername'] + '://' + DB['username'] +':' + DB['password'] + '@' + DB['servername'] + ':' + DB['port'] + '/' + DB['database'] + '?' + 'driver=' + DB['driver']
# print(conn_string)

#  mssql+pyodbc://nebojsa:tras659@VERDI:1433/SANiN?driver=SQL Server Native Client 11.0
    
# Kreiranje konekcije sa MSSQL
engine = create_engine(
    DB['drivername'] 
    + '://' 
    + DB['username'] 
    +':'
    + DB['password'] 
    + '@' 
    + DB['servername'] 
    + ':' 
    + DB['port']
    + '/' 
    + DB['database'] 
    + '?' 
    + 'driver=' 
    + DB['driver'])
conn = engine.connect()

# Required for querying tables
metadata = MetaData(conn)

df1 = pd.read_sql_query("[dbo].[mp_scada] 'SCADA_2021'", engine)
df1.head()

df2 = pd.read_sql_query("[dbo].[mp_scada_rez] 'SCADA_2021'", engine)
df2.head()

df3 = pd.read_sql_query("[dbo].[mp_scada_p] 'SCADA_2021'", engine)
df3.head()

TableName = 'SCADA_2021'

df1.to_csv(TableName+'.csv', index=False)

df2.to_csv(TableName+'_REZ.csv', index=False)

df3.to_csv(TableName+'_p.csv', index=False)

print('Done')

print(df1.columns) 
print("")
print(df2.columns) 
print("")
print(df3.columns) 

# provere
print(type(df1.vreme))
print(type(df1.vreme[0]))

print(type(df1['I-1-LJB'])) 
print(type(df1['I-1-LJB'][0]))

df1.set_index("vreme", inplace=True)
df1

df2.set_index("vreme", inplace=True)
df2

df3.set_index("vreme", inplace=True)
df3

# Close connection
conn.close()

print('GOTOVO!')
