==> notebook/2 - MZ SCADA (IZVORIŠTA) Bokeh 2020.ipynb <==
import sys

import pandas as pd
import bokeh
from bokeh.plotting import figure
from bokeh.io import output_file, output_notebook, show
from bokeh.models import DatetimeTickFormatter, BoxSelectTool, HoverTool, Label, LabelSet
from bokeh.layouts import widgetbox
from bokeh.models.widgets import Dropdown

print('Python v' + sys.version)
print('Pandas v' + pd.__version__)
print('Bokeh v' + bokeh.__version__)

# priprema tipova kolona

dtypes = {'vreme': 'str',
          'I-1-LJB':'float',
          'I-BBV-2':'float',
          'BBV_II-8':'float',
          'I-2-DIV':'float',
          'I-3-MOK':'float',
          'I-4-KR1':'float',
          'I-4-KR2':'float',
          'I-V-6':'float',
          'M-20-992-13':'float',
          'M-20-992-27':'float',
          'M-170-992-34':'float',
          'M-190-21-35':'float',
          'MRB':'float', 
          'III-8':'float',
          'M-70-81-17':'float',
          'I-5-STU':'float',
          'I-6-MLJ':'float',
          'I-7-M10':'float',
          'I-7-M11':'float',
          'I-7-M2':'float',
          'M-20-30-14':'float',
          'M-20-30-22':'float',
          'M-20-40-18':'float'
         }

parse_dates = ['vreme']
df = pd.read_csv("..\scada_data\SCADA_2020.csv", sep=',', header=0, dtype=dtypes, parse_dates=parse_dates, infer_datetime_format=True)
df.set_index("vreme", inplace=True)

# čitanje kreiranog CSV fajla (SCADA_2020.csv)

df = pd.read_csv("..\scada_data\SCADA_2020.csv", sep=',', header=0, dtype=dtypes, parse_dates=parse_dates, infer_datetime_format=True)

# određivanje indeks kolone
df.set_index("vreme", inplace=True)

# definisanje kolona
vreme = df.index
LJB = df['I-1-LJB']          # 1
BNI = df['I-BBV-2']          # 2
BAB = df['BBV_II-8']         # 3
DIV = df['I-2-DIV']          # 4
MOK = df['I-3-MOK']          # 5
KR1 = df['I-4-KR1']          # 6
KR2 = df['I-4-KR2']          # 7
VIN = df['I-V-6']            # 8
VN1 = df['M-20-992-13']      # 9
VN2 = df['M-20-992-27']      # 10
VN3 = df['M-170-992-34']     # 11
VN4 = df['M-190-21-35']      # 12

M10 = df['I-7-M10']          # 13
M11 = df['I-7-M11']          # 14
M20 = df['I-7-M2']           # 15

STU = df['I-5-STU']          # 16
MRB = df['MRB']              # 17
MET = df['III-8']            # 18
BUB = df['M-70-81-17']       # 19

MLJ = df['I-6-MLJ']          # 20

KNJ = df['M-20-30-14']       # 21
DJM = df['M-20-30-22']       # 22

NTX = df['M-10-20-1']        # 23
ROD = df['M-10-40-3']        # 24
LDL = df['M-20-40-18']       # 25

# print(df.columns) 
# print(df.index[1])
# print(df['I-1-LJB'][0], type(df['I-1-LJB'][0]))

# postavke za bokeh dijagram

TOOLS = ["xpan", "xwheel_zoom", "box_zoom", "reset", "save"]

# TOOLTIPS = [
#     ("MP", "$name"),
#     ("(Q,t)", "($y, $x)")
# ]


hover = HoverTool(tooltips=[
    ("MP", "$name"),
    ("Q", "$y{0.0 a}"),
    ("datum", "$x{%F}"),
    ("vreme", "$x{%H:%M}")
#   ( 'close',  '$@{adj close}{%0.2f}' ), # use @{ } for field names with spaces
#   ( 'volume', '@volume{0.00 a}'      ),
    ],

    formatters={
        "$x": 'datetime'      #, use 'datetime' formatter for '@date' field
#       '@{adj close}':'printf', # use 'printf' formatter for '@{adj close}' field
                                 # use default 'numeral' formatter for other fields
    },

    # display a tooltip whenever the cursor is vertically in line with a glyph
    mode='mouse'
)

f = figure(
           plot_width=1500,
           plot_height=500,
           x_axis_type='datetime',
           y_range=(-200, 1200),
           tools=TOOLS+[hover],
#            tooltips=TOOLTIPS,
           active_scroll='xwheel_zoom',
           background_fill_color="#f9f9f9",
           toolbar_location="above"
          )     

f.add_tools(BoxSelectTool(dimensions="width"))

f.title.text_color="blue"
f.title.text_font="arial"
f.title.text_font_style="bold"
f.title.text_font_size="14pt" # "1.5em" 
f.title.text_alpha=1 #floating point between 0 (transparent) and 1 (opaque)

f.xaxis.minor_tick_line_color=None
f.yaxis.minor_tick_line_color=None
f.xaxis.axis_label="Време"
f.yaxis.axis_label="Проток (лит/сек)" 

f.xaxis.formatter=DatetimeTickFormatter(hours=["%H:%M"],days=["%d %B"],months=["%B %Y"],years=["%Y"])

f.toolbar.autohide = True

# krive merača protoka

# ИЗВОРИШТА

# f.title.text="Мерачи Протока НИВОС - Изворишта"

# f.line(vreme, LJB, color="brown", name="Љуберађа", legend_label = 'Љуберађа')
# f.line(vreme, DIV, color="red", name="Дивљана", legend_label = 'Дивљана')
# f.line(vreme, MOK, color="purple", name="Мокра", legend_label = 'Мокра')
# f.line(vreme, KR1, color="orange", name="Крупац (стара ПС)", legend_label = 'Крупац (стара ПС)')
# f.line(vreme, KR2, color="yellow", name="Крупац (нова ПС)", legend_label = 'Крупац (нова ПС)')
# f.line(vreme, STU, color="cyan", name="Студена", legend_label = 'Студена')
# f.line(vreme, MLJ, color="green", name="Миљковац", legend_label = 'Миљковац')
# f.line(vreme, M10, color="blue", name="Медијана 1", legend_label = 'Медијана 1')
# f.line(vreme, M11, color="skyblue", name="Медијана (потисни)", legend_label = 'Медијана (потисни)')
# f.line(vreme, M20, color="powderblue", name="Медијана 2", legend_label = 'Медијана 2')

# Зонски мерачи протока

# РЕЗЕРВОАР ВИНИК

# f.line(vreme, VIN, color="blue")
# f.line(vreme, VN1, color="green")
# f.line(vreme, VN2, color="lightseagreen")
# f.line(vreme, VN3, color="cyan")
# f.line(vreme, VN4, color="lightsteelblue")


# f.line(vreme, BNI, color="red")
# f.line(vreme, BAB, color="purple")

# Мерна Зона 30

# f.title.text="Мерна Зона 30"

# f.line(vreme, M20, color="powderblue", name="Медијана 2")
# f.line(vreme, DJM, color="red", name="Ђачко острво (Медијана)")
# f.line(vreme, KNJ, color="green", name="Књажевачка")

### Кидање АЦ 500 23-октобар-2020

labela =Label(x=70, y=70, x_units='screen', text='LABELA', render_mode='css', 
              border_line_color='black', border_line_alpha=1.0,
              background_fill_color='white', background_fill_alpha=1.0)
f.add_layout(labela)

f.title.text="Оштећење цевовода АЦ ∅500 - 23. Октобар 2020"

f.line(vreme, VIN, line_width = 1, color="blue", name="Виник (улаз)", legend_label = 'Виник (улаз)')
f.line(vreme, VN1, line_width = 2, color="powderblue", name="Виник 1 (подвожњак)", legend_label = 'Виник 1 (подвожњак)')
f.line(vreme, NTX, line_width = 1, color="red", name="Нитекс", legend_label = 'Нитекс')
f.line(vreme, ROD, line_width = 1, color="green", name="Рода (Меркатор)", legend_label = 'Рода (Меркатор)')
f.line(vreme, LDL, line_width = 1, color="black", name="Lidl", legend_label = 'Lidl')

f.legend.location = 'top_left'
f.legend.title = 'Мерач протока'
f.legend.title_text_font = 'Arial'
f.legend.title_text_font_size = '10pt'

show(f)

# labela =Label(x=70, y=70, x_units='screen', text='LABELA', render_mode='css', 
#               border_line_color='black', border_line_alpha=1.0,
#               background_fill_color='white', background_fill_alpha=1.0)
# f.add_layout(labela)

# f.title.text="Оштећење цевовода АЦ ∅500 - 23. Октобар 2020"

# f.line(vreme, VN1, line_width = 2, color="powderblue", name="Виник 1 (подвожњак)", legend_label = 'Виник 1 (подвожњак)')
# f.line(vreme, NTX, line_width = 1, color="red", name="Нитекс", legend_label = 'Нитекс')
# f.line(vreme, ROD, line_width = 1, color="green", name="Рода (Меркатор)", legend_label = 'Рода (Меркатор)')

# f.legend.location = 'top_left'
# f.legend.title = 'Мерач протока'
# f.legend.title_text_font = 'Arial'
# f.legend.title_text_font_size = '10pt'


# menu = [("Option 1", "item_1"), ("Option 2", "item_2")]
# dropdown = Dropdown(label="Dropdown Menu", button_type="warning", menu=menu)
